BL-5872 move sign language api methods to own class
