+++ src/BloomExe/BloomExe.csproj
@@ -619,6 +619,7 @@
     <Compile Include="web\controllers\BookSettingsApi.cs" />
     <Compile Include="web\controllers\BrandingApi.cs" />
     <Compile Include="web\BloomHttpListenerContext.cs" />
+    <Compile Include="web\controllers\SignLanguageApi.cs" />
     <Compile Include="web\IBloomWebSocketServer.cs" />
     <Compile Include="web\controllers\MusicApi.cs" />
     <Compile Include="web\controllers\ToolboxApi.cs" />
